Refactor tags input and use it in the plugin options
